Rm SetIndent in json.NewEncoder / JSON file name
This patch removes the SetIndent function in the
helper func NewJSONEncoder as it broke downstream tests.
This patch also collapses the files back into json.go and
json_test.go.
